service creation fucntion moved out of main file
