+++ packages/__init__.py
@@ -2,3 +2,4 @@ from .file_actions import *
 from .master_class import *
 from .time import *
 from .ipy import *
+from .convert_service import *
